==> 3-secondary-analysis/R/0_config.R <==

#-------------------------------------------------------------------------------
# @Organization - Proctor Foundation - UCSF
# @Project - WASHB-SEP-Season 
# @Author - Pearl Ante-Testard, pearl.ante@ucsf.edu
# @Description - Configuration for maps
#-------------------------------------------------------------------------------

library(haven)
library(here)
library(mgcv)
library(tidymv)
library(tidygam)
library(sf) # Mapping tool
library(tidyverse) # Syntax for data interaction
library(viridis) # Color palette
library(stars) # Read .tif file
library(ggrepel) # Graph label position
library(colorspace)
