==> 3-secondary-analysis/R/2_popdensity_under3y.Rmd <==
---
title: "Mapping children < 3y in Bangladesh"
author: "Pearl Ante-Testard"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: default
    highlight: default
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
knit: (function(inputFile, encoding) {
        rmarkdown::render(inputFile, 
                          encoding   = encoding, 
                          output_dir = here::here("3-secondary-analysis/output")
                          )})
---

```{r clean environment, include=FALSE}

rm(list=ls(all=TRUE))

```


```{r configuration, message=FALSE}

library(here)
source(here::here("3-secondary-analysis/R", "0_config.R"))

```

```{r read worldpop tiff, message=FALSE}

# Aggregated total adult population per grid-cell in Bangladesh
agg_tiff <- read_stars(here::here("1-data", "0-worldpop",
                      "bgd_ppp_2014_1km_Aggregated.tif"))

# Transform into sf object
agg_tiff_sf=st_as_sf(agg_tiff)

# shapefile
bgd_shp_adm2 <- st_read(here::here("1-data", "0-gadm",
                    "gadm41_BGD_2.json"))  

# WASH Benefits Bangladesh clusters
df_gps <- read_dta(file = here::here("1-data", "0-untouched",
                                     "6. WASHB_Baseline_gps.dta"))

df_publicid <- read_csv(file = here::here("1-data", "0-untouched",
                                          "public-ids.csv"))

# individual datasets included in the analysis
df_analysis_svy12 <- readRDS(file = here::here("1-data", "2-final",
                                               "enrol_diar_tr_wealth_indiv_svy012.rds")) %>%
                                                filter(svy!=0)

```

```{r estimate 3y}

# Estimate the under-3 population
agg_tiff_sf <- agg_tiff_sf %>%
  mutate(
         pop_0_14 = bgd_ppp_2014_1km_Aggregated.tif*(0.31), 
         # proportion of 0.31 was based on the World Bank etsimate for under 14: https://data.worldbank.org/indicator/SP.POP.0014.TO.ZS?locations=BD&view=chart
         pop_3andunder = pop_0_14*(3/14)
         ,popdensity = ntile(pop_3andunder, 10)
         ) 

```


```{r washb clusters}

df_gps_publicid <- left_join(df_publicid, df_gps, by = "dataid")
df_gps_publicid <- df_gps_publicid[,-c(1:3)]
colnames(df_gps_publicid) <- c("clusterid","dataid","block","qgpslong","qgpslat","entrydate")

df_gps_publicid_washb_sepseason <- left_join(df_analysis_svy12, df_gps_publicid, by = c("dataid",
                                                                                         "clusterid",
                                                                                         "block"))
df_gps_publicid_clust <- df_gps_publicid_washb_sepseason %>%
  st_as_sf(coords = c("qgpslong", "qgpslat"), crs = 4326) %>%
  group_by(clusterid) %>%
  dplyr::summarise() %>%
  st_cast("POLYGON")

df_gps_publicid_block <- df_gps_publicid_washb_sepseason %>%
  st_as_sf(coords = c("qgpslong", "qgpslat"), crs = 4326) %>%
  group_by(block) %>%
  dplyr::summarise() %>%
  st_cast("POLYGON")

```


```{r pop density plot}

# Colors
pal <- hcl.colors(10, "plasma", rev = TRUE, alpha = 0.7)

# Plot Density
bgd_density <- ggplot() +
  geom_sf(data = agg_tiff_sf, 
          aes(fill = round(popdensity,0)), 
          color=NA,
          show.legend=TRUE) + 
  scale_fill_gradientn(colors = pal) +
  geom_sf(data = bgd_shp_adm2, fill = NA) +
  geom_point(data=df_gps_publicid_clust, size=0.3, shape = 21, fill = "lightgray",
            aes(geometry=geometry), stat = "sf_coordinates") +
  coord_sf() + 
  theme_void() +
  theme(legend.position = "bottom",text = element_text(size=7.5)) +
 labs(
    fill = "Population density of \nchildren < 3y per grid-cell \nat a resolution of 1km \nowith WASH Benefits \ntrial clusters")                
 
bgd_density 

```

```{r save}

# plot
ggsave(bgd, file = here::here("3-secondary-analysis",
                              "output",
                              "map-totalpop-under3-bgd.png"), height=8,width=7)

ggsave(bgd_density, file = here::here("3-secondary-analysis",
                              "output",
                              "map-popdensity-under3-bgd.png"), height=8,width=7)

# data frame
saveRDS(agg_tiff_sf, file = here::here("1-data",
                              "2-final",
                              "pop-under3-grid-cell-bgd.rds"))

write_csv(agg_tiff_sf, file = here::here("1-data",
                              "2-final",
                              "pop-under3-grid-cell-bgd.csv"))

```

```{r session info}
sessionInfo()

```
