==> 3-secondary-analysis/R/4_diar_prevdiffgam_under3_wealth_monsoon.Rmd <==
---
title: "Estimating diarrhea cases prevented by WASH among children < 3y, indexed by relative wealth rank, per grid-cell and district during the monsoon season"
author: "Pearl Ante-Testard"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: default
    highlight: default
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
knit: (function(inputFile, encoding) {
        rmarkdown::render(inputFile, 
                          encoding   = encoding, 
                          output_dir = here::here("3-secondary-analysis/output")
                          )})
---

Here, diarrhea cases prevented by combined WASH per wealth rank were calculated by:

excess diarrhea cases of control vs intervention = proportion of children > 3 years multiplied by the prevalence difference indexed by wealth rank during rainy season

GAM estimates were based on the WASH Benefits Bangladesh cluster randomized trial. We used these GAM estimates to extrapolate to the whole country of Bangladesh, while masking for urban areas since the WASHB trial were conducted in rural areas in Gazipur, Kishoreganj, Mymensingh and Tangail districts.


```{r clean environment, echo=FALSE}

rm(list=ls(all=TRUE))

```

```{r configuration, message=FALSE}

library(here)
library(cowplot)
library(sp)
library(raster)
library(scales)
library(ggpubr)
library(patchwork)

source(here::here("3-secondary-analysis/R", "0_config.R"))

```

```{r read data,  message=F}

# world population in 2014 (from worldpop)
pop_worldpop_bgd <- readRDS(file = here::here("1-data",
                            "2-final",
                            "pop-under3-grid-cell-bgd.rds")) 
st_crs(pop_worldpop_bgd) # WGS 84

# wealth index layer in 2011 (from worldpop)
wealth_worldpop_bgd <- readRDS(file = here::here("1-data",
                              "2-final",
                              "wealthindex-grid-cell-bgd.rds")) 
st_crs(wealth_worldpop_bgd) # WGS 84

# diarrhea prevalence difference between control and interv arms from GAM (WASHB trial)
diar_prevdifference_wealth <- readRDS(file = here::here("1-data","2-final",
                           "gam-diar-prevdifference-interv-wealthrank-rainy.RDS"))
# shapefiles (from GADM)
BGD_shp_adm0 <- st_read(here::here("1-data", "0-gadm",
                   "gadm41_BGD_0.json"))
st_crs(BGD_shp_adm0) # WGS 84
#BGD_shp_adm0 <- st_transform(BGD_shp_adm0, "+proj=longlat +datum=WGS84 +no_defs")
#st_crs(BGD_shp_adm0) # +proj=longlat +datum=WGS84 +no_defs 
BGD_shp_adm2 <- st_read(here::here("1-data", "0-gadm",
                    "gadm41_BGD_2.json")) 
st_crs(BGD_shp_adm2) # WGS 84
#BGD_shp_adm2 <- st_transform(BGD_shp_adm2, "+proj=longlat +datum=WGS84 +no_defs")
#st_crs(BGD_shp_adm2) # +proj=longlat +datum=WGS84 +no_defs
#BGD_shp_adm2 <- st_transform(bgd_shp_adm2, "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs") 

# urban raster layer (from Francois based on the Global Human Settlement)
ras <- readRDS(here::here("1-data", "2-final", "Urban_BGD"))
st_crs(ras) # +proj=longlat +datum=WGS84 +no_defs 
#ras <- projectRaster(ras, crs = crs(BGD_shp_adm2))
st_crs(ras)
extent(ras)
#class      : Extent 
#xmin       : 85.84792 
#xmax       : 95.19292 
#ymin       : 20.64474 
#ymax       : 26.73082 
extent(BGD_shp_adm2)
#class      : Extent 
#xmin       : 88.0113 
#xmax       : 92.6737 
#ymin       : 20.7411 
#ymax       : 26.6341 

# WASH Benefits Bangladesh clusters
df_gps <- read_dta(file = here::here("1-data", "0-untouched",
                                     "6. WASHB_Baseline_gps.dta"))
st_crs(df_gps) # NA

df_publicid <- read_csv(file = here::here("1-data", "0-untouched",
                                          "public-ids.csv"))
st_crs(df_publicid) # NA

# individual datasets included in the analysis
df_analysis_svy12 <- readRDS(file = here::here("1-data", "2-final",
                                               "enrol_diar_tr_wealth_indiv_svy012.rds")) %>%
                                                filter(svy!=0)

```

```{r wealth rank, message=F}

# rounding wealth rank to 1 decimal point before merging to other datasets
wealth_worldpop_bgd <- wealth_worldpop_bgd %>% 
                      mutate(wealth_rank=round(wealth_rank, 1)) 
st_crs(wealth_worldpop_bgd) # WGS 84
#wealth_worldpop_bgd <- st_transform(wealth_worldpop_bgd, "+proj=longlat +datum=WGS84 +no_defs")
#st_crs(wealth_worldpop_bgd) # +proj=longlat +datum=WGS84 +no_defs

diar_prevdifference_wealth <- diar_prevdifference_wealth %>% 
                      mutate(wealth_rank=round(wealth_rank, 1)) 
st_crs(diar_prevdifference_wealth) # NA
```

```{r process data,  message=F}

## join worldpop population 2014 and worldpop wealth 2011
wealth_popunder3_gps <- st_join(wealth_worldpop_bgd, pop_worldpop_bgd, #, left=T
                                ) %>%
                        filter(!is.na(bgd2011wipov.tif))
wealth_popunder3_gps <-wealth_popunder3_gps[!duplicated(wealth_popunder3_gps$bgd_ppp_2014_1km_Aggregated.tif), ]
st_crs(wealth_popunder3_gps) # WGS 84

## join with GAM estimates 
diar_wealth_pop_gps_prevdifference <- left_join(wealth_popunder3_gps %>% as.data.frame(), 
                                          diar_prevdifference_wealth,
                                          by="wealth_rank")
diar_wealth_pop_gps_prevdifference_sf <- st_as_sf(diar_wealth_pop_gps_prevdifference)
diar_wealth_pop_gps_prevdifference_sf <-diar_wealth_pop_gps_prevdifference_sf[
  !duplicated(diar_wealth_pop_gps_prevdifference_sf$bgd_ppp_2014_1km_Aggregated.tif), ]
st_crs(diar_wealth_pop_gps_prevdifference) # NA

diar_wealth_pop_gps_prevdifference_sf <- diar_wealth_pop_gps_prevdifference_sf %>%
                            mutate(excess_diar=pop_3andunder*difference) #%>%
                           # st_as_sf(coords = "geometry", crs = 4326)
st_crs(diar_wealth_pop_gps_prevdifference_sf) #  WGS 84

```


```{r masking urban areas,  message=F}

## transforming urban/rural layer to dataframe
r.spdf <- as(ras, "SpatialPixelsDataFrame")
st_crs(r.spdf) # +proj=longlat +datum=WGS84 +no_defs 
r.df <- as.data.frame(r.spdf)
head(r.df)

# urban (needed for the urban map)
r.df.urban <- subset(r.df, layer==1)
st_crs(r.df.urban) # NA

## transforming into points
r.df.sf <- r.df %>%
  st_as_sf(coords = c("x","y")
           ) 

# set CRS for the points to be the same as shapefile
st_crs(r.df.sf) <- st_crs(BGD_shp_adm2)
#r.df.sf_adm2 <- st_join(r.df.sf, st_make_valid(BGD_shp_adm2)) 

#ggplot() +
 # geom_sf(data = BGD_shp_adm2) +
 # geom_sf(data = r.df.sf_adm2) +
 # theme_minimal()

# some points are out of the polygon, lets filter out
# so we get only points in the polygon
points_sf_joined <- 
  st_join(diar_wealth_pop_gps_prevdifference_sf, r.df.sf) %>%  # spatial join to get intersection of points and poly
  filter(!is.na(layer)) # layer just one col from the polygon data that I chose to filter on, could use any. The idea is to get only the points that fall in the polygon

# plot again 
ggplot() +
  geom_sf(data = BGD_shp_adm2) +
  geom_sf(data = points_sf_joined) +
  theme_minimal()

st_crs(points_sf_joined)

## removing NAs in excess_diar
joined.r <- points_sf_joined %>%
  filter(!is.na(excess_diar))
st_crs(joined.r) # +proj=longlat +datum=WGS84 +no_defs  
#joined.r$excess_diar[joined.r$layer == 1] <- NA
#joined.r$excess_diar[joined.r$layer == "NA"] <- NA

## join the shapefile with the layers for the district estimates
BGD_shp_adm2_valid <- st_make_valid(BGD_shp_adm2)
joined <- st_join(BGD_shp_adm2_valid, joined.r) %>%  # spatial join to get intersection of points and poly
  filter(!is.na(layer)) %>%
  filter(!is.na(excess_diar))
#st_crs(joined) # +proj=longlat +datum=WGS84 +no_defs 

#ggplot() +
#  geom_sf(data = bgd_shp_adm2) +
#  geom_sf(data = joined) +
#  theme_minimal()

```



```{r and washb clusters,  message=F}

df_gps_publicid <- left_join(df_publicid, df_gps, by = "dataid")
df_gps_publicid <- df_gps_publicid[,-c(1:3)]
colnames(df_gps_publicid) <- c("clusterid","dataid","block","qgpslong","qgpslat","entrydate")

df_gps_publicid$dataid <- as.character(df_gps_publicid$dataid)
df_gps_publicid$clusterid <- as.character(df_gps_publicid$clusterid)
df_analysis_svy12$dataid <- as.character(df_analysis_svy12$dataid)
df_analysis_svy12$clusterid <- as.character(df_analysis_svy12$clusterid)
df_gps_publicid_washb_sepseason <- left_join(df_analysis_svy12, df_gps_publicid, by = c("dataid",
                                                                                         "clusterid",
                                                                                         "block"))
df_gps_publicid_clust <- df_gps_publicid_washb_sepseason %>%
  st_as_sf(coords = c("qgpslong", "qgpslat"), crs = 4326) %>%
  group_by(clusterid) %>%
  dplyr::summarise() %>%
  st_cast("POLYGON")

df_gps_publicid_block <- df_gps_publicid_washb_sepseason %>%
  st_as_sf(coords = c("qgpslong", "qgpslat"), crs = 4326) %>%
  group_by(block) %>%
  dplyr::summarise() %>%
  st_cast("POLYGON")

```

# WASH Benefits Bangladesh clusters

```{r plot clusters, message=F}

map_bgd_wash_clust <- ggplot(df_gps_publicid_clust) +
  geom_sf(data = BGD_shp_adm0, fill = "grey") +
  #geom_sf(color="blue") +
  geom_point(size=0.5, shape = 21, aes(geometry=geometry),
    stat = "sf_coordinates") +
  geom_sf(data = BGD_shp_adm2, fill = NA) +
  theme_void() +
  #theme(legend.position = "bottom") +
  ggtitle("WASH Benefits Bangladesh trial clusters included in the analysis") +
  theme(plot.title = element_text(hjust = 0.5))
 
map_bgd_wash_clust
# the trial was located in Gazipur, Mymensingh and Tangail districts 

```


# WASH Benefits Bangladesh block

```{r plot block,  message=F}

map_bgd_wash_block <- ggplot(df_gps_publicid_block) +
  geom_sf(data = BGD_shp_adm0, fill = "grey") +
  #geom_sf(color="blue") +
  geom_point(size=1, shape = 21, aes(geometry=geometry),
    stat = "sf_coordinates") +
  geom_sf(data = BGD_shp_adm2, fill = NA) +
  theme_void() +
  #theme(legend.position = "bottom") +
  ggtitle("WASH Benefits Bangladesh trial blocks included in the analysis") +
  theme(plot.title = element_text(hjust = 0.5))
 
map_bgd_wash_block
# the trial was located in Gazipur, Kishoreganj, Mymensingh and Tangail districts 

```

# Population density and WASH Benefits Study clusters

```{r pop density plot, message=F}

# Colors
pal <- hcl.colors(10, "mako", rev = TRUE)

# Plot Density
bgd_density <- ggplot() +
  geom_sf(data = pop_worldpop_bgd, 
          aes(fill = popdensity), 
          color=NA,
          show.legend=TRUE) + 
  scale_fill_gradientn(colors=pal,
                       values=rescale(seq(0,10,by=2)),
                       limits=c(0,10), 
                       #n.breaks = 8
                       breaks = c(0,2,4,6,8,10)) +
  geom_point(data=df_gps_publicid_clust, size=0, shape = 21, color = "#E34D34" ,
            aes(geometry=geometry), stat = "sf_coordinates") +
  geom_sf(data = BGD_shp_adm2, fill = NA, color="black") +
  coord_sf() + 
  theme_void() +
  theme(legend.position = "right",text = element_text(size=6)) +
 labs(
    fill = "Population density \nof children < 3y \nper grid-cell")                
 
bgd_density <- bgd_density + annotate(
    geom = "curve", x=91, y=25.9, xend = 90.5, yend = 25, 
    curvature = .1, arrow = arrow(length = unit(1.5, "mm")), col="#E34D34"
  ) +
  annotate("text", x=91.3, y=25.9, label= "WASH Benefits Study clusters \n(n=360)",
           col="#E34D34", size=1.3, parse=F)

bgd_density

```

# Wealth index layer 

```{r wealth, message=F}

# Read tif file for wealth index, 2011
wipov_tif=read_stars(here::here("1-data", "0-worldpop",
                                "bgd2011wipov.tif"))
wipov_sf=st_as_sf(wipov_tif)

# Wealth rank
wipov_sf <- wipov_sf %>%
  mutate(wealth_rank = rank(bgd2011wipov.tif)/nrow(.),
         wealth_rank = round(wealth_rank,1))

# Colors
pal <- hcl.colors(10, "Rocket", rev = T)

bgd_wealth_cont <- ggplot() +
  geom_sf(data = wipov_sf, 
          aes(fill = wealth_rank), 
          color=NA,
          show.legend=TRUE, alpha=1) + 
  scale_fill_gradientn(colors=pal,
                       values=rescale(seq(0,1,by=0.2)),
                       limits=c(0,1), 
                       #n.breaks = 8
                       breaks = seq(0,1,by=0.2)) +
  geom_sf(data=BGD_shp_adm2, fill=NA, color="black") +
  coord_sf() + 
  theme_void()+
  theme(legend.position = "right",text = element_text(size=6)) +
  labs(
    fill = "Relative rank based \non the wealth index")
bgd_wealth_cont

#ggsave(bgd_wealth_cont, file = here::here("3-secondary-analysis", 
                                        #  "output",
                                        #  "map-wealthrank-bgd.png"), height=8,width=7)

```

# Masked urban layer

The urban/rural layer from Francois Rerolle was based on the Global Human Settlement.

```{r plot per grid cell masked, message=F}

# urban/rural layer from Francois
#plot_ras <- plot(ras)

plot_maskedurban <- ggplot() +  
   geom_sf(data=BGD_shp_adm2, color="black", fill="grey") +
   geom_tile(data=r.df.urban, aes(x=x,y=y), fill="darkblue") +
   coord_sf() + 
   theme_void() +
   labs(subtitle = "Urban areas that were masked in \nthe spatial analysis (blue color)") +
   #ggtitle("Urban areas (blue color) that were masked in the spatial analysis") +
   theme(text = element_text(size=7.5),  plot.caption = element_text(hjust = 0.5))
  
plot_maskedurban
```

## Diarrhea cases prevented for children <3y per grid-cell at a 1 km res. 

```{r diarrhea cases prevented masked, message=F}

#joined.r$excess_diar[joined.r$layer == 1] <- NA
 #Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    
  # 0.00    1.18    2.07    2.69    3.58   27.31

#clipped_data <- st_intersection(joined.r, BGD_shp_adm2)

#colnames(joined.r) <- c("layer","bgd_ppp_2014_1km_Aggregated","pop_0_14","pop_3andunder","popdensity",
                       # "bgd2011wipov","wealth_rank","wealthdensity","block","dummy","difference","SE","pvalue",
                       # "CI_upper","CI_lower","sig_diff","group","excess_diar",
                       # "geometry")

joined.r$excess_diar[joined.r$pop_3andunder <= 2] <- NA
#joined.r$layer[is.na(joined.r$layer)] = 1
joined.r$excess_diar[joined.r$layer == 1] <- NA
#joined.r$excess_diar[joined.r$layer == 2] <- NA
# layer
#    0      1 
#111519  26870 

#joined.r_2 <- joined.r[,-c(22)]

#summary(joined.r_2$excess_diar)

joined.r_2 <- joined.r %>%
  mutate(excess_diar_monthly = round(as.numeric(excess_diar)*4,0)) %>%
#  > table(joined.r_2$excess_diar_monthly)
#   0     1     2     3     4     5     6     7     8     9    10    11    12    13    14    15    16    17    18 
# 1015 10079  4632  6634 11282 13897 13341 11356  8980  6896  5035  3834  2850  2068  1434   963   746   497   362 
#   19    20    21    22    23    24    25    26    27    28    29    30    31    32    33    34    35    36    37 
#  245   182   143   106    65    71    53    41    27    33    30    18    10    20    16    15     5     6    12 
#   38    39    40    41    43    44    45    46    47    48    49    50    54    56    57    59    61    64    70 
#    8     4     1     3     2     3     1     5     2     1     3     4     2     2     1     1     3     1     1 
#   82   106 
#    1     1 
  mutate(excess_diar_monthly_recoded =                                
                                      ifelse(excess_diar_monthly>=32,32,
                                      ifelse(excess_diar_monthly==31,31,
                                      ifelse(excess_diar_monthly==30,30,
                                      ifelse(excess_diar_monthly==29,29,
                                      ifelse(excess_diar_monthly==28,28,
                                      ifelse(excess_diar_monthly==27,27,
                                      ifelse(excess_diar_monthly==26,26,
                                      ifelse(excess_diar_monthly==25,25,
                                      ifelse(excess_diar_monthly==24,24,
                                      ifelse(excess_diar_monthly==23,23,
                                      ifelse(excess_diar_monthly==22,22,
                                      ifelse(excess_diar_monthly==21,21,
                                      ifelse(excess_diar_monthly==20,20,
                                      ifelse(excess_diar_monthly==19,19,
                                      ifelse(excess_diar_monthly==18,18,
                                      ifelse(excess_diar_monthly==17,17,
                                      ifelse(excess_diar_monthly==16,16,
                                      ifelse(excess_diar_monthly==15,15,
                                      ifelse(excess_diar_monthly==14,14,
                                      ifelse(excess_diar_monthly==13,13,
                                      ifelse(excess_diar_monthly==12,12,
                                      ifelse(excess_diar_monthly==11,11,
                                      ifelse(excess_diar_monthly==10,10, 
                                      ifelse(excess_diar_monthly==9,9, 
                                      ifelse(excess_diar_monthly==8,8, 
                                      ifelse(excess_diar_monthly==7,7, 
                                      ifelse(excess_diar_monthly==6,6, 
                                      ifelse(excess_diar_monthly==5,5, 
                                      ifelse(excess_diar_monthly==4,4, 
                                      ifelse(excess_diar_monthly==3,3, 
                                      ifelse(excess_diar_monthly==2,2, 
                                      ifelse(excess_diar_monthly==1,1, 
                                      ifelse(excess_diar_monthly==0,0, NA
                                             ))))))))))))))))))))))))))))))))))

table(joined.r_2$excess_diar_monthly_recoded)


## plot the diarrhea cases prevented by WASH
pal <- hcl.colors(7, "Plasma", rev = T)

plot_diar_averted_bgd_gridcell_rural <- ggplot() +
  #geom_tile(data=joined_rural_2, aes(x=x,y=y), fill=NA) +
  #geom_sf(data=joined, fill=NA, color="black") +
  geom_sf(data = joined.r_2, 
          aes(color=excess_diar_monthly_recoded),
          #color=NA,
          show.legend=TRUE) +
  scale_color_gradientn(colors=pal,
                       values=rescale(seq(0,28,by=4)),
                       limits=c(0,28), 
                       n.breaks = 8,
                       breaks = seq(0,28,by=4),
                       labels = c("0","4","8","12","16","20","24",
                                  "28+"#,
                                  ),
                       na.value = 'darkgrey'
  ) +
  scale_y_continuous(breaks=seq(0,28,by=4)) +
  #geom_tile(data=joined_urban.r, aes(x=x,y=y), fill="grey") +
  geom_sf(data=BGD_shp_adm2, fill=NA, color="black") +
  coord_sf() + 
  #new_scale_fill() +
  theme_void()+
  theme(legend.position = "right",text = element_text(size=6)) +
  labs(
    color = "Diarrhea cases \nprevented in \nchildren <3y per \nmonth per grid-cell \nduring monsoon")

plot_diar_averted_bgd_gridcell_rural <- plot_diar_averted_bgd_gridcell_rural + 
  annotate(
    geom = "curve",  x=91.3, y=21.6, xend = 91.8, yend = 22, 
    curvature = .3, arrow = arrow(length = unit(1.75, "mm")), col="darkgrey"
  ) +
  annotate("text", x=91.2, y=21.3, label= "Masked urban areas \nand grid-cell with 1-2 \nchildren < 3y \n(excluded from estimates)",
           col="darkgrey", size=1.3, parse=F)

plot_diar_averted_bgd_gridcell_rural

#ggsave(plot_diar_averted_bgd_gridcell_rural, file = here::here("3-secondary-analysis", 
                                        #  "output",
                                        #  "map-diar-under3-wealthrank-bgd-rural-gridcell-2.jpg"), height=8,width=7)

#0,4,8,12,16,20,24,28,32
```

## Diarrhea cases prevented per 1000 children < 3y per week in each district 

```{r plot per per district masked, message=F}

## subset rural layers by district
#joined.r_2_cast <- st_cast(joined.r_2, to="MULTIPOLYGON")

joined_grouped_rural <- subset(joined, layer==0) 
joined_grouped_rural$excess_diar[joined_grouped_rural$pop_3andunder <= 2] <- NA

joined_grouped_rural <- joined_grouped_rural %>%
  group_by(NAME_2) %>%
  mutate(excess_diar_sum = sum(excess_diar, na.rm = T),
         excess_diar_per1000under3 = excess_diar_sum/1000,
         excess_diar_per1000under3_monthly = excess_diar_per1000under3*4,
         # Calculate weights inversely proportional to standard errors
         #weights = pop_3andunder / sum(pop_3andunder),
         # Combine standard errors
         #combined_se = sqrt(sum(weights * SE^2) /sum (weights)),
         #combined_se = mean(SE),
         #combined_se = sqrt(sum(excess_diar^2) / (length(excess_diar) - 1)),
         combined_se = sum(sqrt(SE^2*pop_3andunder)),
         #combined_se = sum(sqrt(SE^2*pop_3andunder)),
         ci_upper = excess_diar_sum+1.96*combined_se,
         ci_upper = (ci_upper/1000)*4,
         ci_lower = excess_diar_sum-1.96*combined_se,
         ci_lower = (ci_lower/1000)*4
         )

#> summary(joined_grouped_rural$excess_diar_per1000under3_monthly)
#  Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#  1.310   7.556  10.071  11.478  14.218  37.708 

## plot the diarrhea cases prevented by WASH excluding urban areas
#pal <- hcl.colors(5, "Viridis", rev = F, alpha = 0.9)
#my_breaks = c(0, 1, 2, 3, 4)
joined_grouped_rural <-joined_grouped_rural[
  !duplicated(joined_grouped_rural$NAME_2), ]

#> summary(joined_grouped_rural$excess_diar_per1000under3_monthly)
#Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
# 0.8401  6.8443  9.1519 10.3808 13.1319 34.5760

plot_diar_averted_bgd_rural_dist <- ggplot() +
  #geom_sf(data=BGD_shp_adm2, fill=NA, color="black") +
  geom_sf(data = joined_grouped_rural, 
          aes(fill=excess_diar_per1000under3_monthly),
          #color=NA,
          show.legend=TRUE) +
  #scale_fill_gradientn(colors = pal, guide="legend") +
  scale_fill_viridis_c(alpha = 1,begin = 0,end = 1,aesthetics = "fill",direction = -1,
                       values=scales::rescale(seq(0,40,by=4)),
                       breaks = seq(0,40,by=4),limits=c(0,40)
                       )+
  #scale_y_continuous(breaks=seq(0,20,by=4)) +
  geom_sf(data=BGD_shp_adm2, fill=NA, color="black") +
  coord_sf() + 
  #new_scale_fill() +
  theme_void()+
  theme(legend.position = "none",text = element_text(size=7.5)) #+
  #labs(
    #fill = "Diarrhea cases prevented \nper 1000 children < 3y per \nmonth in each district \n(masked urban areas)")
plot_diar_averted_bgd_rural_dist

plot_diar_averted_bgd_rural_dist_print <- ggplot() +
  geom_sf(data = joined_grouped_rural, 
          aes(fill=excess_diar_per1000under3_monthly),
          color=NA,
          show.legend=TRUE) +
  #scale_fill_gradientn(colors = pal, guide="legend") +
  scale_fill_viridis_c(alpha = 1,begin = 0,end = 1,aesthetics = "fill",direction = -1,
                       values=scales::rescale(seq(0,40,by=4)),
                       breaks = seq(0,40,by=4),
                       limits=c(0,40)
                       )+
  #scale_y_continuous(breaks=seq(0,18,by=2)) +
  geom_sf(data=BGD_shp_adm2, fill=NA, color="black") +
  coord_sf() + 
  #new_scale_fill() +
  theme_void()+
  theme(legend.position = "right",text = element_text(size=6)) +
  labs(
    fill = "Diarrhea cases prevented \nper 1000 children < 3y per \nmonth in each district \nduring monsoon")

plot_diar_averted_bgd_rural_dist_print 

#ggsave(plot_diar_averted_bgd_rural_dist, file = here::here("3-secondary-analysis", 
                                         # "output",
                                         # "map-diarprevented-maskedurban-bgd-2.png"), height=8,width=7)


```

## Diarrhea cases prevented per 1000 children < 3y per week at country level

```{r plot per district masked, message=F}

## subset rural layers by district
#joined_grouped_rural_adm0 <- subset(joined, layer==0) 
#joined_grouped_rural_adm0$excess_diar[joined_grouped_rural_adm0$pop_3andunder <= 2] <- NA

joined_grouped_rural_adm0 <- joined_grouped_rural %>%
  group_by(COUNTRY) %>%
  mutate(excess_diar_sum_dist = sum(excess_diar_sum, na.rm = T),
         excess_diar_per1000under3 = excess_diar_sum_dist/1000,
         excess_diar_per1000under3_monthly = excess_diar_per1000under3*4,
         total_pop_3andunder = sum(pop_3andunder, na.rm = T),
         combined_se_nat = sum(sqrt(combined_se^2*pop_3andunder)),
         #combined_se = sum(sqrt(SE^2*pop_3andunder)),
         ci_upper = excess_diar_sum_dist+1.96*combined_se_nat,
         ci_upper = (ci_upper/1000)*4,
         ci_lower = excess_diar_sum_dist-1.96*combined_se_nat,
         ci_lower = (ci_lower/1000)*4)

joined_grouped_rural_adm0$excess_diar_sum_dist
joined_grouped_rural_adm0$excess_diar_per1000under3
table(joined_grouped_rural_adm0$excess_diar_per1000under3_monthly,joined_grouped_rural_adm0$NAME_1)
#table(joined_grouped_rural_adm0$excess_diar_per1000under3_monthly,joined_grouped_rural_adm0$NAME_2)

```

Barplot

```{r barplot no urban areas, message=F}

## barplot of the diarrhea cases prevented per district
#joined_grouped_rural <-joined_grouped_rural[
  #!duplicated(joined_grouped_rural$NAME_2), ]

joined_grouped_rural <- joined_grouped_rural %>%
  mutate(NAME_2 = factor(NAME_2,levels=joined_grouped_rural$NAME_2[order(joined_grouped_rural$excess_diar_per1000under3)]))

table(joined_grouped_rural$excess_diar_per1000under3_monthly)
summary(joined_grouped_rural$excess_diar_per1000under3_monthly)
# Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#  1.310   7.556  10.071  11.478  14.218  37.708 

options(scipen=999)

#pal <- hcl.colors(5, "Viridis", rev = F, alpha = 0.9)
barplot_diar_averted_bgd_rural_dist <- ggplot(joined_grouped_rural, aes(x=NAME_2, y=excess_diar_per1000under3_monthly, fill=excess_diar_per1000under3*4)) +
    geom_bar(stat = "identity") +
    scale_fill_viridis_c(alpha = 1,begin = 0,end = 1,aesthetics = "fill",direction = -1,
                       values=scales::rescale(seq(0,40,by=4)),
                       #breaks = seq(0,18,by=2),
                       limits=c(0,40)
                       )+
   scale_y_continuous(breaks=seq(0,40,by=4)) +
   geom_hline(yintercept=seq(0,40,by=4), colour="white") +
   geom_errorbar( aes(x=NAME_2, ymin=ci_lower, ymax=ci_upper), width=0.4, colour="darkred", alpha=0.9, linewidth=0.5)+
    coord_flip() +
    xlab(" ") +
    #scale_y_continuous(position = 'right') +
    ylab(
      "Diarrhea cases prevented per 1000 children < 3y \nper month in each district during monsoon") +
    theme_bw() +
    theme(text = element_text(size=6)) +
    theme(legend.position = "none") +
    theme(panel.grid.major=element_blank(), panel.grid.minor=element_blank()) 

barplot_diar_averted_bgd_rural_dist

#ggsave(barplot_diar_averted_bgd_rural_dist, file = here::here("3-secondary-analysis", 
                                      #    "output",
                                       #   "barplot-diarprevented-maskedurban-bgd-2.png"), height=8,width=7)

```

# Combined plots 

```{r merge plots, message=F}


map_district_composite <- barplot_diar_averted_bgd_rural_dist + inset_element(plot_diar_averted_bgd_rural_dist, left = 0.4, bottom = 0, right = 1, top = 0.9)

map_district_composite

plot_composite <- ggarrange(bgd_density, bgd_wealth_cont, plot_diar_averted_bgd_gridcell_rural, 
          map_district_composite, #+ rremove("x.text"), 
          labels = c("A", "B", "C", "D"),
          ncol = 2, nrow = 2, align = "hv", widths = c(3.5, 3.5))
plot_composite


ggsave(plot_composite, file = here::here("3-secondary-analysis", 
                                          "output",
                                          "map-composite-bgd-29-gam-wealth-monsoon.pdf"), 
                                          height=190, width=180, units="mm")

```


```{r}
sessionInfo()

```
